fix(nvim): use a non-integer table to store attached language tools
A lua list was used to keep track of attached language tools (LSP,
formatters, linters). This caused a problem if any of the above
restarted: it would trigger LspAttach which adds to the list. The result
was the tool appearing multiple times. Using the tool name as the index
avoids this. It is essentially a very poor set.