feat(wezterm): try out Victor Mono to replace Iosevka
I still like Iosevka, but it doesn't load efficiently in Wezterm.
Specifically, the brew formula updated to include one large font file,
instead of the individual ones. For whatever reason, wezterm has trouble
handling the larger file, and it results in taking multiple seconds to
load a new window.

Victor Mono is still a condensed (but not as condensed) font, so we'll
try it out.